Add Scala.js support to core and integration modules
