+++ core/src/main/scalajs/sttp/tapir/TapirExtensions.scala
@@ -1,0 +1,3 @@
+package sttp.tapir
+
+trait TapirExtensions
