+++ core/src/main/scalajs/sttp/tapir/internal/UrlencodedData.scala
@@ -1,0 +1,26 @@
+package sttp.tapir.internal
+
+import java.nio.charset.Charset
+import scala.scalajs.js.URIUtils
+
+private[tapir] object UrlencodedData {
+  def decode(s: String, charset: Charset): Seq[(String, String)] = {
+    def decodeSpace(in: String): String = in.replace('+', ' ')
+    s.split("&")
+      .toList
+      .flatMap(kv =>
+        kv.split("=", 2) match {
+          case Array(k, v) =>
+            Some((URIUtils.decodeURIComponent(decodeSpace(k)), URIUtils.decodeURIComponent(decodeSpace(v))))
+          case _ => None
+        }
+      )
+  }
+
+  def encode(s: Seq[(String, String)], charset: Charset): String = {
+    def encodeSpace(in: String): String = in.replace("%20", "+")
+    s.map { case (k, v) =>
+      s"${encodeSpace(URIUtils.encodeURIComponent(k))}=${encodeSpace(URIUtils.encodeURIComponent(v))}"
+    }.mkString("&")
+  }
+}
