+++ core/src/main/scalajvm/sttp/tapir/TapirExtensions.scala
@@ -1,0 +1,9 @@
+package sttp.tapir
+
+import java.io.File
+import java.nio.file.Path
+
+trait TapirExtensions {
+  def fileBody: EndpointIO.Body[File, File] = rawBinaryBody[File]
+  def pathBody: EndpointIO.Body[File, Path] = binaryBody[File, Path]
+}
