+++ core/src/test/scala/sttp/tapir/generic/FormCodecDerivationTest.scala
@@ -4,7 +4,6 @@ import java.math.{BigDecimal => JBigDecimal}
 
 import com.github.ghik.silencer.silent
 import sttp.tapir.SchemaType.{SObjectInfo, SProduct}
-import sttp.tapir.util.CompileUtil
 import sttp.tapir.{Codec, CodecFormat, DecodeResult, FieldName, Schema, Validator, encodedName}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -56,17 +55,6 @@ class FormCodecDerivationTest extends AnyFlatSpec with Matchers {
 
     codec.decode("f1=v1&f2=10") shouldBe DecodeResult.Value(Test4(Some("v1"), 10))
     codec.decode("f2=10") shouldBe DecodeResult.Value(Test4(None, 10))
-  }
-
-  it should "report a user-friendly error when a codec for a parameter cannot be found" in {
-    val error = CompileUtil.interceptEval("""
-        |import sttp.tapir._
-        |trait NoCodecForThisTrait
-        |case class Test5(f1: String, f2: NoCodecForThisTrait)
-        |implicitly[Codec[String, Test5, CodecFormat.XWwwFormUrlencoded]]
-        |""".stripMargin)
-
-    error.message should include("Cannot find a codec between types: List[String] and NoCodecForThisTrait")
   }
 
   it should "use the right schema for a two-arg case class" in {
@@ -190,7 +178,17 @@ class FormCodecDerivationTest extends AnyFlatSpec with Matchers {
     val codec = implicitly[Codec[String, Test1, CodecFormat.XWwwFormUrlencoded]]
 
     // when
-    codec.encode(test1) shouldBe "f1=str&f2=42&f3=228&f4=322&f5=69&f6=27.0&f7=33.0&f8=true&f9=1337.7331&f10=31337.73313"
+    codec
+      .encode(test1)
+      .shouldBe(
+        // In Scala.js in general, a trailing .0 is omitted i.e. 1.0.toString == "1", instead of "1.0"
+        // See https://www.scala-js.org/doc/semantics.html
+        if (System.getProperty("java.vm.name") == "Scala.js")
+          "f1=str&f2=42&f3=228&f4=322&f5=69&f6=27&f7=33&f8=true&f9=1337.7331&f10=31337.73313"
+        else
+          "f1=str&f2=42&f3=228&f4=322&f5=69&f6=27.0&f7=33.0&f8=true&f9=1337.7331&f10=31337.73313"
+      )
+
     codec.decode("f1=str&f2=42&f3=228&f4=322&f5=69&f6=27.0&f7=33.0&f8=true&f9=1337.7331&f10=31337.73313") shouldBe DecodeResult.Value(test1)
   }
