+++ core/src/test/scala/sttp/tapir/generic/MultipartCodecDerivationTest.scala
@@ -1,11 +1,8 @@
 package sttp.tapir.generic
 
-import java.io.File
-
 import com.github.ghik.silencer.silent
 import sttp.model.{Header, MediaType, Part}
 import sttp.tapir.SchemaType._
-import sttp.tapir.util.CompileUtil
 import sttp.tapir.{DecodeResult, FieldName, MultipartCodec, RawPart, Schema, Validator, encodedName}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -49,17 +46,6 @@ class MultipartCodecDerivationTest extends AnyFlatSpec with Matchers {
     codec.decode(createStringParts(List(("f2", "10")))) shouldBe DecodeResult.Value(Test4(None, 10))
   }
 
-  it should "report a user-friendly error when a codec for a parameter cannot be found" in {
-    val error = CompileUtil.interceptEval("""
-                                            |import sttp.tapir._
-                                            |trait NoCodecForThisTrait
-                                            |case class Test5(f1: String, f2: NoCodecForThisTrait)
-                                            |implicitly[MultipartCodec[Test5]]
-                                            |""".stripMargin)
-
-    error.message should include("Cannot find a codec between a List[T] for some basic type T and: NoCodecForThisTrait")
-  }
-
   it should "use the right schema for a case class with part metadata" in {
     // given
     case class Test6(f1: String, f2: Int)
@@ -70,20 +56,6 @@ class MultipartCodecDerivationTest extends AnyFlatSpec with Matchers {
       SProduct(
         SObjectInfo("sttp.tapir.generic.MultipartCodecDerivationTest.<local MultipartCodecDerivationTest>.Test6"),
         List((FieldName("f1"), implicitly[Schema[String]]), (FieldName("f2"), implicitly[Schema[Int]]))
-      )
-    )
-  }
-
-  it should "use the right schema for a two-arg case class" in {
-    // given
-    case class Test1(f1: Part[File], f2: Int)
-    val codec = implicitly[MultipartCodec[Test1]].codec
-
-    // when
-    codec.schema.map(_.schemaType) shouldBe Some(
-      SProduct(
-        SObjectInfo("sttp.tapir.generic.MultipartCodecDerivationTest.<local MultipartCodecDerivationTest>.Test1"),
-        List((FieldName("f1"), implicitly[Schema[File]]), (FieldName("f2"), implicitly[Schema[Int]]))
       )
     )
   }
@@ -141,45 +113,6 @@ class MultipartCodecDerivationTest extends AnyFlatSpec with Matchers {
         )
       )
     )
-  }
-
-  it should "generate a codec for a case class with file part" in {
-    // given
-    case class Test1(f1: File)
-    val codec = implicitly[MultipartCodec[Test1]].codec
-    val f = File.createTempFile("tapir", "test")
-
-    try {
-      // when
-      codec.encode(Test1(f)) shouldBe List(Part("f1", f, fileName = Some(f.getName), contentType = Some(MediaType.ApplicationOctetStream)))
-      codec.decode(List(Part("f1", f, fileName = Some(f.getName)))) shouldBe DecodeResult.Value(Test1(f))
-    } finally {
-      f.delete()
-    }
-  }
-
-  it should "use the right schema for an optional file part with metadata 2" in {
-    // given
-    case class Test1(f1: Part[Option[File]], f2: Int)
-    val codec = implicitly[MultipartCodec[Test1]].codec
-    val f = File.createTempFile("tapir", "test")
-
-    // when
-    try {
-      // when
-      codec.encode(Test1(Part("?", Some(f), otherDispositionParams = Map("a1" -> "b1")), 12)) shouldBe List(
-        Part("f1", f, otherDispositionParams = Map("a1" -> "b1"), contentType = Some(MediaType.ApplicationOctetStream)),
-        Part("f2", "12", contentType = Some(MediaType.TextPlain))
-      )
-      codec.decode(List(Part("f1", f, fileName = Some(f.getName)), Part("f2", "12"))) shouldBe DecodeResult.Value(
-        Test1(Part("f1", Some(f), fileName = Some(f.getName)), 12)
-      )
-    } finally {
-      f.delete()
-    }
-
-    codec.encode(Test1(Part("f1", None), 12)) shouldBe List(Part("f2", "12", contentType = Some(MediaType.TextPlain)))
-    codec.decode(List(Part("f2", "12"))) shouldBe DecodeResult.Value(Test1(Part("f1", None), 12))
   }
 
   it should "generate a codec for a one-arg case class with implicit validator" in {
