+++ json/playjson/src/test/scala/sttp/tapir/json/play/TapirJsonPlayTests.scala
@@ -5,7 +5,6 @@ import play.api.libs.json._
 import sttp.tapir._
 import sttp.tapir.DecodeResult._
 
-import java.util.Date
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
@@ -49,23 +48,6 @@ class TapirJsonPlayTests extends AnyFlatSpec with Matchers {
 
   it should "encode and decode Long type" in {
     testEncodeDecode(1566150331L)
-  }
-
-  it should "encode and decode using custom Date serializer" in {
-    val d = new Date
-    testEncodeDecode(d)
-  }
-
-  it should "Fail to encode a badly formatted date" in {
-    val codec = TapirJsonPlayCodec.readsWritesCodec[Date]
-    val encoded = "\"OOPS-10-10 11:20:49.029\""
-
-    codec.decode(encoded) match {
-      case _: DecodeResult.Failure =>
-        succeed
-      case Value(d) =>
-        fail(s"Should not have been able to decode this date: $d")
-    }
   }
 
   it should "encode to non-prettified Json" in {
