+++ project/plugins.sbt
@@ -7,3 +7,4 @@ addSbtPlugin("com.typesafe.sbt" % "sbt-twirl" % "1.5.0")
 addSbtPlugin("org.scalameta" % "sbt-mdoc" % "2.2.11")
 addSbtPlugin("com.eed3si9n" % "sbt-projectmatrix" % "0.6.0")
 addSbtPlugin("org.jetbrains" % "sbt-ide-settings" % "1.1.0")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.3.0")
